requirements.txt
requests==2.32.3
pytest==8.2.1
python-dotenv==1.0.1
allure-pytest==2.13.5
pytest-xdist==3.6.1
jsonschema==4.23.0